refactor(security): move verifyEnvelopeSignature to its call site, document MSP trust boundary
Addresses review feedback on #1611:

- verifyEnvelopeSignature moved from gossip.go into client.go, its only
  caller (peersForChannel/endorser), per reviewer suggestion.
- Documents that verifyEnvelopeSignature is a self-consistency check
  only: it proves a gossip envelope was signed by the key in its own
  self-claimed identity, not that the identity is a CA-issued,
  MSP-trusted peer. mspx509.NewIdentityFromBytes is the same
  no-CA-binding lightweight provider already documented on
  msp/x509.Provider.DeserializeVerifier, so this check inherits that
  exact limitation. Tracked for follow-up hardening in #1623.
