+++ platform/fabric/core/generic/discovery/gossip.go
@@ -11,7 +11,6 @@ import (
 
 	"github.com/hyperledger-labs/fabric-smart-client/pkg/utils/errors"
 	"github.com/hyperledger-labs/fabric-smart-client/pkg/utils/proto"
-	mspx509 "github.com/hyperledger-labs/fabric-smart-client/platform/fabric/core/generic/msp/x509"
 )
 
 // SignedGossipMessage contains a GossipMessage and the Envelope from which it
@@ -68,44 +67,4 @@ func EnvelopeToGossipMessage(e *gossip.Envelope) (*SignedGossipMessage, error) {
 		GossipMessage: msg,
 		Envelope:      e,
 	}, nil
-}
-
-// verifyEnvelopeSignature checks that e.Signature is a valid signature over
-// e.Payload produced by the private key corresponding to the given
-// serialized identity. This binds a relayed gossip envelope to the specific
-// identity it is shipped alongside in the same discovery response, so an
-// envelope whose signature does not verify against that identity's key -
-// including a forged/garbage signature, or one produced by a different key -
-// is rejected instead of being trusted unconditionally.
-//
-// allowEmptySignature must be true only for AliveMessage/MembershipInfo
-// envelopes. A real Fabric peer signs its own self-referential AliveMessage
-// with protoext.NoopSign (gossip/discovery's Self()), which produces a nil
-// Signature by design, and the Discovery service exposes that self-entry
-// as-is (discovery/support/gossip's Peers(), via SelfMembershipInfo()). This
-// is inherent to how real Fabric peers answer Discovery queries - even
-// upstream Fabric's own discovery client performs no verification at all on
-// these envelopes - so rejecting a nil signature here would fail every
-// Discovery query in which the responding peer is itself among the reported
-// peers/endorsers, which is the common case in small networks. StateInfo
-// self-entries do not share this exemption: real peers always produce a
-// genuine signature for them, even for themselves (see
-// gossipChannel.setupSignedStateInfoMessage, which calls a real signer, not
-// NoopSign), so callers must pass false for StateInfo envelopes to keep that
-// verification strict.
-func verifyEnvelopeSignature(e *gossip.Envelope, identity []byte, allowEmptySignature bool) error {
-	if e == nil {
-		return errors.New("nil envelope")
-	}
-	if allowEmptySignature && len(e.Signature) == 0 {
-		return nil
-	}
-	_, verifier, err := mspx509.NewIdentityFromBytes(identity)
-	if err != nil {
-		return errors.Wrap(err, "failed deriving verifier from peer identity")
-	}
-	if err := verifier.Verify(e.Payload, e.Signature); err != nil {
-		return errors.Wrap(err, "signature does not verify against claimed identity")
-	}
-	return nil
 }
